+++ react/CHANGELOG.md
@@ -18,6 +18,7 @@ sidebar_custom_props: { 'icon': '📰' }
 ## Unreleased
 
 -   💥 **Breaking Change**: This project now requires THEOplayer version 7.0.0 or higher. ([#60](https://github.com/THEOplayer/web-ui/pull/60))
+-   🚀 Added `<THEOliveDefaultUI>` that provides a default UI for THEOlive streams. ([#58](https://github.com/THEOplayer/web-ui/pull/58))
 
 ## v1.7.2 (2024-03-18)